life_tables

Complete cohort life tables data as provided Stats NZ

life_tables

Format

A data frame with nine variables:

yearofbirth

Year of birth

age

age in years

sex

Sex

percentile

percentile of the estimate

dx

number of deaths that would occur between integral ages

ex

Expected number of years of life remaining at age x

llx

Average number alive in the age interval out of 100,000 people born

lx

Number alive at exact age out of 100,000 people born

mx

proportion of individuals of age x dying by age x+1

px

Probabilty a person who reaches that age lives for the age interval

sx

Proportion of age group x to x+5 surviving another 5 years

qx

Probabilty a person who reaches that age dies within the age interval